GSAGH-550: further refactor
